--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,59 +1,0 @@
----
-name: Bug report
-about: 버그 제보!
-title: "[\U0001F41E Bug Report]"
-labels: "\U0001F41E BugFix"
-assignees: Kimgyuilli
-
----
-
----
-name: 🐞 Bug Report
-about: 버그를 발견하셨나요? 여기서 제보해주세요.
-title: "[Bug] "
-labels: bug
-assignees: ''
-
----
-
-## 🐛 버그 설명
-
-발생한 버그에 대해 간략하고 명확하게 설명해주세요.
-
-예시:
-> 로그인 시 500 에러가 발생하며, 서버 로그에는 `NullPointerException`이 출력됩니다.
-
----
-
-## 🔁 버그 재현 절차
-
-1. ...
-2. ...
-3. ...
-4. 에러 확인
-
----
-
-## 🤔 기대한 동작
-
-버그가 없을 경우 기대되는 정상 동작을 설명해주세요.
-
----
-
-## 🖥️ 환경 정보
-
-- 운영체제: [예: Windows 11, macOS Sonoma]
-- 브라우저 / 클라이언트: [예: Chrome, Edge, Android 앱]
-- 사용한 버전: [예: Spring Boot 3.1.2, React 18.2.0 등]
-
----
-
-## 📸 첨부파일 (선택사항)
-
-스크린샷, 콘솔 로그, 오류 메시지 등 참고할 수 있는 자료가 있다면 첨부해주세요.
-
----
-
-## 📎 추가 내용
-
-그 외 공유하고 싶은 내용이 있다면 작성해주세요.
