--- .github/ISSUE_TEMPLATE/custom.md
@@ -1,19 +1,0 @@
----
-name: Custom issue template
-about: 템플릿을 새로 만들고 싶다면
-title: "[Example]"
-labels: "\U0001F4C3 Docs, ✅ Test"
-assignees: ''
-
----
-
----	
-name: Hello Issue
-about: Custom Issue Template Test
-title: ''
-labels: ''
-assignees: ''
-
----
-
-👋 Custom Issue Template Test
