--- .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,42 +1,0 @@
----
-name: Feature request
-about: 새로운 Feature 제안
-title: "[✨ Feature]"
-labels: "✨ Feature"
-assignees: Kimgyuilli
-
----
-
----
-name: ✨ Feature Request
-about: 새로운 기능이 필요하신가요? 아이디어를 공유해주세요!
-title: "[Feature] "
-labels: ✨ Feature
-assignees: '담당자 이름름'
-
----
-
-## 🌟 제안할 기능
-
-어떤 기능을 제안하시는지 간결하고 명확하게 설명해주세요.
-
-예시:
-> 게시글에 좋아요 기능을 추가하면 사용자 참여도가 올라갈 수 있습니다.
-
----
-
-## 📌 기능의 필요성
-
-이 기능이 왜 필요한지, 어떤 문제를 해결할 수 있는지 설명해주세요.
-
----
-
-## 🧩 구현 아이디어 (선택사항)
-
-어떻게 구현하면 좋을지 간단한 아이디어가 있다면 공유해주세요.
-
----
-
-## 📎 추가 내용
-
-그 외 참고할 만한 자료, 디자인, 예시 등이 있다면 첨부해주세요.
