--- .github/ISSUE_TEMPLATE/question.md
@@ -1,37 +1,0 @@
----
-name: Question
-about: 질문 있어요
-title: "[QnA]"
-labels: "\U0001F64B‍♂️ Question"
-assignees: Kimgyuilli
-
----
-
----
-name: ❓ Question / Help
-about: 궁금한 점이나 도움이 필요한 내용을 올려주세요!
-title: "[QnA] "
-labels: question
-assignees: ''
-
----
-
-## ❓ 질문 내용
-
-궁금한 점을 간결하고 명확하게 작성해주세요.
-
-예시:
-> 서비스 구조에서 프론트와 백엔드 간의 인증 흐름이 잘 이해되지 않습니다. 설명 부탁드립니다.
-
----
-
-## 📄 관련 코드 또는 상황
-
-관련된 코드, API, 기능 위치 등을 작성해주세요. (선택사항)
-
-```java
-// 예시 코드 블럭
-@PostMapping("/login")
-public ResponseEntity<?> login(@RequestBody LoginRequest request) {
-    ...
-}
