--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,25 +1,0 @@
-## 개요
-<!---- 변경 사항 및 관련 이슈에 대해 간단하게 작성해주세요. 어떻게보다 무엇을 왜 수정했는지 설명해주세요. -->
-
-<!---- Resolves: #(Isuue Number) -->
-
-## PR 유형
-어떤 변경 사항이 있나요?
-
-- [ ] 새로운 기능 추가
-- [ ] 버그 수정
-- [ ] CSS 등 사용자 UI 디자인 변경
-- [ ] 코드에 영향을 주지 않는 변경사항(오타 수정, 탭 사이즈 변경, 변수명 변경)
-- [ ] 코드 리팩토링
-- [ ] 주석 추가 및 수정
-- [ ] 문서 수정
-- [ ] 테스트 추가, 테스트 리팩토링
-- [ ] 빌드 부분 혹은 패키지 매니저 수정
-- [ ] 파일 혹은 폴더명 수정
-- [ ] 파일 혹은 폴더 삭제
-
-## PR Checklist
-PR이 다음 요구 사항을 충족하는지 확인하세요.
-
-- [ ] 커밋 메시지 컨벤션에 맞게 작성했습니다.  Commit message convention 참고  (Ctrl + 클릭하세요.) 
-- [ ] 변경 사항에 대한 테스트를 했습니다.(버그 수정/기능에 대한 테스트).
